time added and conversion to UTC
